Don't disable audio processing prematurely
Multiple WebRTCVoiceMediaChannels can be associated with a single
AudioProcessing module.  Disabling some of the audio processing
when a single WebRTCVoiceMediaChannel's streams mute is incorrect,
as it may disable audio processing for the other
WebRTCVoiceMediaChannels.

For now, don't disable any of the audio processing.  This may use
more, but it won't be much.